feat: Add collab close dlc protocol record
